refactor(metal): move the two submission timing instruments out of gpu.rs
Pure move, no behaviour change: mm_timing_add (FERROX_METAL_MM_TIMING)
and gpu_timing_note (FERROX_METAL_GPU_TIMING) go to timing.rs, along
the seam GitHub issue #149 touches. gpu.rs 8901 -> 8789.
