+++ crates/ferrox-metal/src/dispatch.rs
@@ -109,6 +109,7 @@ mod tests {
             ("lib.rs", include_str!("lib.rs")),
             ("mem_ranges.rs", include_str!("mem_ranges.rs")),
             ("moe_ids.rs", include_str!("moe_ids.rs")),
+            ("timing.rs", include_str!("timing.rs")),
         ];
         // This file holds the wrapper (the one legitimate caller of the
         // raw dispatch method) and the patterns this test searches for,
